fix: record trash ui
